Renamed and split network info implementation
