+++ src_features/provide_network_info/cmd_network_info.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <stdint.h>
+
+uint16_t handle_network_info(uint8_t p1,
+                             uint8_t p2,
+                             const uint8_t *data,
+                             uint8_t length,
+                             unsigned int *tx);
